backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
django-import-export==3.3.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.6
MarkupPy==1.14
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
